requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
wikipedia-api>=0.6.0

# Database
neo4j>=5.12.0
python-dotenv>=1.0.0

# LLM integration
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
networkx>=3.2.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.66.0